% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap for confidence intervals}
\usage{
bootstrap(param_hat, K, W, alpha, max_t)
}
\arguments{
\item{param_hat}{parameters c(beta, gamma, I0W, phi) on the log(it) scale}

\item{K}{the flight connectivity matrix within China}

\item{W}{the flight connectivity matrix outside China}

\item{alpha}{the latent period}

\item{max_t}{time to simulate up to}
}
\value{
a data.frame of bootstrapped parameter samples
}
\description{
Bootstrap for confidence intervals
}
