% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{import_matrix}
\alias{import_matrix}
\title{Import contact matrix}
\usage{
import_matrix(filename, replace_NA_w_zero = FALSE, ...)
}
\arguments{
\item{filename}{the filename containing the matrix, csv with first column as row.names}

\item{replace_NA_w_zero}{replace NAs with zero}
}
\value{
a matrix
}
\description{
Import contact matrix
}
