% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Likelihood function for Wuhan nCoV-2019}
\usage{
LogLikelihood(
  y,
  z,
  N,
  K,
  W,
  sim_fun,
  phi_mask = (rownames(K) == "Wuhan"),
  agg_up_to = 11
)
}
\arguments{
\item{y}{a \eqn{n \times T} matrix of case reports in China}

\item{z}{a \eqn{m \times T} matrix of case reports elsewhere}

\item{N}{the population sizes within China, length n}

\item{K}{the within-China air travel matrix, \eqn{n \times n}}

\item{W}{the international air travel matrix, \eqn{n \times m}}

\item{sim_fun}{a function which returns a simulation from a disease model}

\item{phi_mask}{a vector of 0s and 1s determining which cities in China to apply the
underreporting parameter \eqn{phi} to.}

\item{agg_up_to}{aggregate the first however many case detection records, allowing for
a delay in receiving counts from the first few cases.}
}
\value{
a function to calculate the log likelihood.  This function has signature
\code{logp_fn(params, visualise=FALSE)} where \code{params} is a vector of parameters
\code{c(beta, gamma, I0W, phi)}.  If \code{visualise} is \code{TRUE}, then parameter values
are printed to the console and a graph showing how the ODE mean function matches the observed
timeseries (in Wuhan) is displayed.  This is useful for tracking the progression of various
optimisers.
}
\description{
This function returns a function encoding the likelihood for the nCoV-2019
model.  It assumes Poisson-distributed increments of case
reports in China, and Poisson-distributed increments in numbers of
infected passengers on planes elsewhere.
}
\details{
This function returns a closure -- another function that encapsulates the
data passed to the containing function.  See the return value for the function signature.
}
