% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{NetworkODEModel}
\alias{NetworkODEModel}
\title{Network SEIR metapopulation model}
\usage{
NetworkODEModel(N, K, init_loc, alpha, max_t)
}
\arguments{
\item{N}{the population sizes}

\item{K}{a contact matrix}

\item{init_loc}{the name of the initial location}

\item{alpha}{the latent period}
}
\value{
list of matrices containing simulated results
}
\description{
Network SEIR metapopulation model
}
