% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary_samples}
\alias{summary_samples}
\title{Summarises samples from bootstrapping}
\usage{
summary_samples(samples)
}
\arguments{
\item{samples}{an \eqn{n \times 4} \code{data.frame} of \code{beta}, \code{gamma}, \code{I0W}, \code{phi}}
}
\value{
table of summary data
}
\description{
Summarises samples from bootstrapping
}
