% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Likelihood function for Wuhan nCoV-2019}
\usage{
LogLikelihood(y, z, N, K, W, sim_fun, agg_up_to = 11)
}
\arguments{
\item{y}{a $n \times T$ matrix of case reports in China}

\item{z}{a $m \times T$ matrix of case reports elsewhere}

\item{N}{the population sizes within China, length n}

\item{K}{the within-China air travel matrix, $n \times n$}

\item{W}{the international air travel matrix, $n \times m$}

\item{sim_fun}{a function which returns a simulation from a disease model}

\item{param}{a vector of model parameters c(beta, gamma, I0W, phi)}

\item{visualise}{if TRUE then print out model parameters and a graph of
modelled case detections in Wuhan.  Useful for following progress of optimisers}
}
\value{
a function to calculate the log likelihood

the log likelihood of the data conditional on the model and parameters
}
\description{
Likelihood function for Wuhan nCoV-2019
}
